+++ oB-recon-configure.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 
 """
+
+
 Writes a configuration file for the oB-recon program. This controls which
 tests are run.
 
@@ -10,9 +12,18 @@ Scan1   = Scan 1 target is set up to do a "quick scan" from nmap on target x.
           In order to do a more intense scan, the options would be nmap -T4 -A -v <ipaddress>,
           this is not a configurable toggle within the oB-recon program.
 
-Scan2   = Example
+Scan2   = Scan 2 enumerates the available ciphers via nmap --script ssl-enum-ciphers -p 443 <target>
 
 Scan3   = ARIN.net lookup for the IP in question.
+
+Scan4   = Example.
+Scan5   = Example.
+Scan6   = Example.
+Scan7   = Example.
+Scan8   = Example.
+Scan9   = Example.
+Scan10  = Example.
+
 """
 
 # Import
@@ -28,7 +39,14 @@ def writeConf():
     #conf['connectionstring-example']    = {'servername': 'server', 'port': '23'}
     conf['scans']                       = {'scan1': 'True',\
                                            'scan2': 'False',\
-                                           'scan3': 'False'}
+                                           'scan3': 'False',\
+                                           'scan4': 'False',\
+                                           'scan5': 'False',\
+                                           'scan6': 'False',\
+                                           'scan7': 'False',\
+                                           'scan8': 'False',\
+                                           'scan9': 'False',\
+                                           'scan10': 'False'}
     with open ('conf.ini', 'w') as configfile:
         conf.write(configfile)
     return