+++ changelog.md
@@ -1,5 +1,13 @@
 ### Change Log
 
+0.5 - 7/25/18
+
+- Added placeholders for `scan1(t)` through `scan10(t)` in both function 
+and `scanselect(s,t)`. Just have to add the code/resources for the specific scans.
+- Added checking for SSL configuration as `scan2(t)`. This runs `nmap
+--script ssl-enum-ciphers -p 443 <target>`.
+- 
+
 0.4 - 7/10/18
 
 - Added nmap quick scan as 'scan1'